Renamed acquisition -> measure.
