+++ spectral/measure/mydaq.py
